Add VID/PIDs for Planck EZ and Glow revision
