+++ keyboards/planck/ez/glow/glow.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "ez.h"
