+++ keyboards/planck/ez/glow/rules.mk
@@ -1,0 +1,1 @@
+RGB_MATRIX_ENABLE = IS31FL3737
