+++ keyboards/planck/ez/readme.md
@@ -14,6 +14,12 @@ Make example for this keyboard (after setting up your build environment):
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
 
+# Planck EZ Glow
+
+For the per key RGB version of this keyboard, you want to use the "glow" subdirectory. For example: 
+
+    make planck/ez/glow:defaultrrr
+
 ## Planck EZ Configuration (from Oryx)
 
 ### Indicator LEDs
